requirements.txt
PyPDF2==3.0.1
python-dotenv==1.0.0
flask
flask_cors
google-cloud-aiplatform
google-generativeai